refactor: nodes and relationship extraction
