+++ cognee/tasks/graph/extract_relationships_from_data.py
@@ -1,10 +1,56 @@
 import asyncio
 from typing import List
-
+from pydantic import BaseModel
+from cognee.infrastructure.llm.get_llm_client import get_llm_client
+from cognee.infrastructure.llm.prompts import render_prompt
 from cognee.modules.chunking.models.DocumentChunk import DocumentChunk
-from cognee.modules.data.extraction.knowledge_graph.extract_content_graph import (
-    extract_content_nodes_and_relationships,
-)
+
+
+class PotentialNodesAndRelationships(BaseModel):
+    """Response model containing lists of potential node names and relationships."""
+
+    nodes: List[str]
+    relationships: List[str]
+
+
+async def extract_content_nodes_and_relationships(
+    content: str, n_rounds: int = 2
+) -> tuple[List[str], List[str]]:
+    """Extracts node names and relationships from content through multiple rounds of analysis."""
+    llm_client = get_llm_client()
+    all_nodes: List[str] = []
+    all_relationships: List[str] = []
+    existing_nodes = set()  # Track existing node names in lowercase
+    existing_relationships = set()  # Track existing relationship names in lowercase
+
+    for round_num in range(n_rounds):
+        context = {
+            "previous_nodes": all_nodes,
+            "previous_relationships": all_relationships,
+            "round_number": round_num + 1,
+            "total_rounds": n_rounds,
+        }
+
+        system_prompt = render_prompt("extract_graph_relationships_prompt.txt", context)
+        response = await llm_client.acreate_structured_output(
+            text_input=content,
+            system_prompt=system_prompt,
+            response_model=PotentialNodesAndRelationships,
+        )
+
+        # Only add new nodes that haven't been seen before
+        for node in response.nodes:
+            if node.lower() not in existing_nodes:
+                all_nodes.append(node)
+                existing_nodes.add(node.lower())
+
+        # Only add new relationships that haven't been seen before
+        for relationship in response.relationships:
+            if relationship.lower() not in existing_relationships:
+                all_relationships.append(relationship)
+                existing_relationships.add(relationship.lower())
+
+    return all_nodes, all_relationships
 
 
 async def extract_relationships_from_data(
